Renamings, shorter, hopefully better understandable
