+++ src/convertmode/uint2ascii.go
@@ -1,4 +1,4 @@
-package convertfunctions
+package convertmode
 
 import (
 	"encoding/binary"
